# optimise some of the bounds checks